Add "Disclaimer" content, add "Tools" to TOC
The content of the disclaimer is meant to inform users that this product
is simply a data exercise and it's meant for fun and as a portfolio
project for us developers!

I will be asking for outside input to make sure the disclaimer is
adequate and also not too over-the-top!